# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personalized Zodiac Book API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * @nullable
 */
export type ZodiacOrderIntention =
  | (typeof ZodiacOrderIntention)[keyof typeof ZodiacOrderIntention]
  | null;

export const ZodiacOrderIntention = {
  gift: "gift",
  myself: "myself",
} as const;

/**
 * @nullable
 */
export type ZodiacOrderGender =
  | (typeof ZodiacOrderGender)[keyof typeof ZodiacOrderGender]
  | null;

export const ZodiacOrderGender = {
  male: "male",
  female: "female",
} as const;

/**
 * @nullable
 */
export type ZodiacOrderSexualOrientation =
  | (typeof ZodiacOrderSexualOrientation)[keyof typeof ZodiacOrderSexualOrientation]
  | null;

export const ZodiacOrderSexualOrientation = {
  straight: "straight",
  gay: "gay",
  bisexual: "bisexual",
  prefer_not_to_say: "prefer_not_to_say",
} as const;

/**
 * @nullable
 */
export type ZodiacOrderRelationshipStatus =
  | (typeof ZodiacOrderRelationshipStatus)[keyof typeof ZodiacOrderRelationshipStatus]
  | null;

export const ZodiacOrderRelationshipStatus = {
  single: "single",
  in_relationship: "in_relationship",
  married: "married",
  divorced: "divorced",
  widowed: "widowed",
  not_seeking: "not_seeking",
} as const;

export type ZodiacOrderStatus =
  (typeof ZodiacOrderStatus)[keyof typeof ZodiacOrderStatus];

export const ZodiacOrderStatus = {
  pending_payment: "pending_payment",
  pending: "pending",
  generating: "generating",
  generated: "generated",
  submitting: "submitting",
  processing: "processing",
  shipped: "shipped",
  failed: "failed",
} as const;

export interface ZodiacOrder {
  id: number;
  fullName: string;
  birthday: string;
  birthTime: string;
  birthLocation: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  intention?: ZodiacOrderIntention;
  /** @nullable */
  gender?: ZodiacOrderGender;
  /** @nullable */
  sexualOrientation?: ZodiacOrderSexualOrientation;
  /** @nullable */
  relationshipStatus?: ZodiacOrderRelationshipStatus;
  status: ZodiacOrderStatus;
  /** @nullable */
  generatedContent?: string | null;
  /** @nullable */
  sunSign?: string | null;
  /** @nullable */
  moonSign?: string | null;
  /** @nullable */
  risingSign?: string | null;
  /** @nullable */
  lifePath?: string | null;
  /** @nullable */
  customAffirmations?: string | null;
  /** @nullable */
  luckyNumbers?: string | null;
  /** @nullable */
  referralCode?: string | null;
  /** @nullable */
  referredBy?: string | null;
  referralCount?: number;
  /** @nullable */
  luluOrderId?: string | null;
  /** @nullable */
  luluStatus?: string | null;
  /** @nullable */
  shippingAddress?: string | null;
  /** @nullable */
  priceUsd?: number | null;
  /** @nullable */
  stripeSessionId?: string | null;
  /** @nullable */
  stripePaymentIntentId?: string | null;
  /** @nullable */
  interiorPdfUrl?: string | null;
  /** @nullable */
  coverPdfUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

/**
 * Who this book is for — a gift for someone else, or for oneself
 */
export type CreateZodiacOrderBodyIntention =
  (typeof CreateZodiacOrderBodyIntention)[keyof typeof CreateZodiacOrderBodyIntention];

export const CreateZodiacOrderBodyIntention = {
  gift: "gift",
  myself: "myself",
} as const;

export type CreateZodiacOrderBodyGender =
  (typeof CreateZodiacOrderBodyGender)[keyof typeof CreateZodiacOrderBodyGender];

export const CreateZodiacOrderBodyGender = {
  male: "male",
  female: "female",
} as const;

/**
 * Sexual orientation — shapes the Relationships chapter
 */
export type CreateZodiacOrderBodySexualOrientation =
  (typeof CreateZodiacOrderBodySexualOrientation)[keyof typeof CreateZodiacOrderBodySexualOrientation];

export const CreateZodiacOrderBodySexualOrientation = {
  straight: "straight",
  gay: "gay",
  bisexual: "bisexual",
  prefer_not_to_say: "prefer_not_to_say",
} as const;

/**
 * Current relationship status — shapes the Relationships chapter
 */
export type CreateZodiacOrderBodyRelationshipStatus =
  (typeof CreateZodiacOrderBodyRelationshipStatus)[keyof typeof CreateZodiacOrderBodyRelationshipStatus];

export const CreateZodiacOrderBodyRelationshipStatus = {
  single: "single",
  in_relationship: "in_relationship",
  married: "married",
  divorced: "divorced",
  widowed: "widowed",
  not_seeking: "not_seeking",
} as const;

export interface CreateZodiacOrderBody {
  fullName: string;
  /** ISO date string (YYYY-MM-DD) */
  birthday: string;
  /** Time of birth (HH:MM) */
  birthTime: string;
  /** City, Country */
  birthLocation: string;
  email?: string;
  /** Who this book is for — a gift for someone else, or for oneself */
  intention?: CreateZodiacOrderBodyIntention;
  gender?: CreateZodiacOrderBodyGender;
  /** Sexual orientation — shapes the Relationships chapter */
  sexualOrientation?: CreateZodiacOrderBodySexualOrientation;
  /** Current relationship status — shapes the Relationships chapter */
  relationshipStatus?: CreateZodiacOrderBodyRelationshipStatus;
  /** Personal affirmations and mantras to weave into the book */
  customAffirmations?: string;
  /** Referral code used by the person who referred this customer */
  referredBy?: string;
  /** Whether the user consented to receive marketing emails */
  marketingOptIn?: boolean;
}

export interface SubmitToLuluBody {
  shippingName: string;
  shippingAddress1: string;
  shippingAddress2?: string;
  shippingCity: string;
  shippingState: string;
  shippingZip: string;
  shippingCountry: string;
  email: string;
}

export interface OrderStats {
  total: number;
  pending: number;
  generating: number;
  generated: number;
  shipped: number;
  recentOrders?: ZodiacOrder[];
}

export interface CreateCheckoutSessionBody {
  orderId: number;
}

export interface CheckoutSessionResponse {
  url: string;
}

export interface ReferralInfo {
  code: string;
  referrerName: string;
  discountPercent: number;
  timesUsed: number;
  sunSign?: string | null;
  moonSign?: string | null;
  risingSign?: string | null;
  lifePath?: string | null;
  luckyNumbers?: string | null;
}

export interface SiteSettings {
  id: number;
  priceUsd: number;
  originalPriceUsd: number;
  /** @nullable */
  coverImageUrl?: string | null;
  /**
   * When set, all transactional emails are also sent to this address
   * @nullable
   */
  testEmailOverride?: string | null;
  /** @nullable */
  emailConfirmSubject?: string | null;
  /** @nullable */
  emailConfirmIntro?: string | null;
  /** @nullable */
  emailBookReadySubject?: string | null;
  /** @nullable */
  emailBookReadyIntro?: string | null;
  /** @nullable */
  emailStuckSubject?: string | null;
  /** @nullable */
  emailStuckIntro?: string | null;
  updatedAt: string;
}

export interface UpdateSiteSettingsBody {
  priceUsd?: number;
  originalPriceUsd?: number;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  testEmailOverride?: string | null;
  /** @nullable */
  emailConfirmSubject?: string | null;
  /** @nullable */
  emailConfirmIntro?: string | null;
  /** @nullable */
  emailBookReadySubject?: string | null;
  /** @nullable */
  emailBookReadyIntro?: string | null;
  /** @nullable */
  emailStuckSubject?: string | null;
  /** @nullable */
  emailStuckIntro?: string | null;
}