# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Personalized Zodiac Book API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stripe/create-checkout-session:
    post:
      operationId: createCheckoutSession
      tags: [zodiac-orders]
      summary: Create a Stripe Checkout session for an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutSessionBody"
      responses:
        "200":
          description: Checkout session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"
        "400":
          description: Invalid input
        "404":
          description: Order not found
...
components:
  schemas:
...
    CreateCheckoutSessionBody:
      type: object
      properties:
        orderId:
          type: integer
      required:
        - orderId
    CheckoutSessionResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url
...