# lib/api-zod/src/generated/types/createCheckoutSessionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personalized Zodiac Book API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCheckoutSessionBody {
  orderId: number;
}